refactor: split SettingsScreen into focused section widgets
Extract eight section widgets from the monolithic settings_screen.dart:
- ProfileSection, AppearanceSection, TimerSection, BellsSection,
  BgMusicSection, TagsSection, DataSection, and shared SectionHeader/SettingsTile

The main settings_screen.dart shrinks from 751 lines to ~55 lines.
No behavior changes — pure structural extraction.

Closes #19
