+++ citta/lib/l10n/app_localizations.dart
@@ -88,7 +88,7 @@ import 'app_localizations_zh.dart';
 /// property.
 abstract class AppLocalizations {
   AppLocalizations(String locale)
-      : localeName = intl.Intl.canonicalizedLocale(locale.toString());
+    : localeName = intl.Intl.canonicalizedLocale(locale.toString());
 
   final String localeName;
 
@@ -111,11 +111,11 @@ abstract class AppLocalizations {
   /// of delegates is preferred or required.
   static const List<LocalizationsDelegate<dynamic>> localizationsDelegates =
       <LocalizationsDelegate<dynamic>>[
-    delegate,
-    GlobalMaterialLocalizations.delegate,
-    GlobalCupertinoLocalizations.delegate,
-    GlobalWidgetsLocalizations.delegate,
-  ];
+        delegate,
+        GlobalMaterialLocalizations.delegate,
+        GlobalCupertinoLocalizations.delegate,
+        GlobalWidgetsLocalizations.delegate,
+      ];
 
   /// A list of this localizations delegate's supported locales.
   static const List<Locale> supportedLocales = <Locale>[
@@ -145,7 +145,7 @@ abstract class AppLocalizations {
     Locale('tcy'),
     Locale('te'),
     Locale('ur'),
-    Locale('zh')
+    Locale('zh'),
   ];
 
   /// No description provided for @actionCancel.
@@ -1012,34 +1012,34 @@ class _AppLocalizationsDelegate
 
   @override
   bool isSupported(Locale locale) => <String>[
-        'ar',
-        'as',
-        'bn',
-        'de',
-        'en',
-        'es',
-        'fr',
-        'gu',
-        'he',
-        'hi',
-        'it',
-        'ja',
-        'kn',
-        'kok',
-        'mai',
-        'ml',
-        'mr',
-        'or',
-        'pa',
-        'pt',
-        'ru',
-        'sa',
-        'ta',
-        'tcy',
-        'te',
-        'ur',
-        'zh'
-      ].contains(locale.languageCode);
+    'ar',
+    'as',
+    'bn',
+    'de',
+    'en',
+    'es',
+    'fr',
+    'gu',
+    'he',
+    'hi',
+    'it',
+    'ja',
+    'kn',
+    'kok',
+    'mai',
+    'ml',
+    'mr',
+    'or',
+    'pa',
+    'pt',
+    'ru',
+    'sa',
+    'ta',
+    'tcy',
+    'te',
+    'ur',
+    'zh',
+  ].contains(locale.languageCode);
 
   @override
   bool shouldReload(_AppLocalizationsDelegate old) => false;
@@ -1105,8 +1105,9 @@ AppLocalizations lookupAppLocalizations(Locale locale) {
   }
 
   throw FlutterError(
-      'AppLocalizations.delegate failed to load unsupported locale "$locale". This is likely '
-      'an issue with the localizations generation tool. Please file an issue '
-      'on GitHub with a reproducible sample app and the gen-l10n configuration '
-      'that was used.');
+    'AppLocalizations.delegate failed to load unsupported locale "$locale". This is likely '
+    'an issue with the localizations generation tool. Please file an issue '
+    'on GitHub with a reproducible sample app and the gen-l10n configuration '
+    'that was used.',
+  );
 }
