+++ citta/lib/screens/settings/settings_widgets.dart
@@ -1,0 +1,49 @@
+import 'package:flutter/material.dart';
+import '../../theme/app_theme.dart';
+
+class SectionHeader extends StatelessWidget {
+  final String title;
+  const SectionHeader({super.key, required this.title});
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.fromLTRB(16, 16, 16, 4),
+      child: Text(
+        title.toUpperCase(),
+        style: const TextStyle(
+          fontSize: 12,
+          fontWeight: FontWeight.w600,
+          color: AppColors.textHint,
+          letterSpacing: 1.5,
+        ),
+      ),
+    );
+  }
+}
+
+class SettingsTile extends StatelessWidget {
+  final String title;
+  final String subtitle;
+  final VoidCallback? onTap;
+  final Widget? trailing;
+
+  const SettingsTile({
+    super.key,
+    required this.title,
+    required this.subtitle,
+    this.onTap,
+    this.trailing,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return ListTile(
+      title: Text(title),
+      subtitle: Text(subtitle),
+      trailing: trailing ??
+          const Icon(Icons.chevron_right, color: AppColors.textHint, size: 20),
+      onTap: onTap,
+    );
+  }
+}
